+++ src/Metrics.php
@@ -56,7 +56,7 @@ final class Metrics implements JsonSerializable
         $start->timestamp <= $end->timestamp || throw new UnableToProfile('The ending snapshot was taken before the starting snapshot.');
 
         return new self(
-            cpuTime: TimeUnit::Millisecond->toNanoseconds(self::calculateCpuTime($start, $end)),
+            cpuTime: DurationUnit::Millisecond->toNanoseconds(self::calculateCpuTime($start, $end)),
             executionTime: $end->executionTime - $start->executionTime,
             memoryUsage: $end->memoryUsage - $start->memoryUsage,
             peakMemoryUsage: $end->peakMemoryUsage - $start->peakMemoryUsage,